Move data model to separate file
