+++ ert_shared/storage/blob_api.py
@@ -1,4 +1,4 @@
-from ert_shared.storage import (
+from ert_shared.storage.model import (
     Observation,
     Realization,
     Ensemble,
