+++ tests/storage/__init__.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import Session
 from ert_shared.storage.rdb_api import RdbApi
 from ert_shared.storage.blob_api import BlobApi
 
-from ert_shared.storage import Entities, Blobs
+from ert_shared.storage.model import Entities, Blobs
 
 
 @pytest.fixture(scope="session")
